feat: add euler vault support with redeem, deposit and tests

+++ contracts/test/MerchantRegistry.t.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.20;
 
-import {Test, console2} from "../lib/forge-std/src/Test.sol";
+import {Test} from "../lib/forge-std/src/Test.sol";
 import {MerchantRegistry} from "../src/MerchantRegistry.sol";
 
 contract MerchantRegistryTest is Test {
@@ -19,10 +19,7 @@ contract MerchantRegistryTest is Test {
     address public feeRecipient;
 
     // Events to test
-    event MerchantRegistered(
-        address indexed merchantId,
-        address indexed merchant
-    );
+    event MerchantRegistered(address indexed merchantId, address indexed merchant);
     event MerchantUpdated(address indexed merchantId);
     event AllowedTokenSet(address token, bool allowed);
 
@@ -95,15 +92,10 @@ contract MerchantRegistryTest is Test {
             feeRecipient
         );
 
-        MerchantRegistry.MerchantConfig memory config = registry.getConfig(
-            merchant1
-        );
+        MerchantRegistry.MerchantConfig memory config = registry.getConfig(merchant1);
         assertEq(config.merchantPayout, merchant1);
         assertEq(config.payoutToken, usdcToken);
-        assertEq(
-            uint8(config.mode),
-            uint8(MerchantRegistry.PayoutMode.USDC_ONLY)
-        );
+        assertEq(uint8(config.mode), uint8(MerchantRegistry.PayoutMode.USDC_ONLY));
         assertEq(config.splitBps, 0);
         assertEq(config.protocolFeeBps, 500);
         assertEq(config.feeRecipient, feeRecipient);
@@ -122,13 +114,8 @@ contract MerchantRegistryTest is Test {
             feeRecipient
         );
 
-        MerchantRegistry.MerchantConfig memory config = registry.getConfig(
-            merchant1
-        );
-        assertEq(
-            uint8(config.mode),
-            uint8(MerchantRegistry.PayoutMode.SHARES_ONLY)
-        );
+        MerchantRegistry.MerchantConfig memory config = registry.getConfig(merchant1);
+        assertEq(uint8(config.mode), uint8(MerchantRegistry.PayoutMode.SHARES_ONLY));
         assertEq(config.protocolFeeBps, 300);
     }
 
@@ -144,9 +131,7 @@ contract MerchantRegistryTest is Test {
             feeRecipient
         );
 
-        MerchantRegistry.MerchantConfig memory config = registry.getConfig(
-            merchant1
-        );
+        MerchantRegistry.MerchantConfig memory config = registry.getConfig(merchant1);
         assertEq(uint8(config.mode), uint8(MerchantRegistry.PayoutMode.SPLIT));
         assertEq(config.splitBps, 7000);
         assertEq(config.protocolFeeBps, 200);
@@ -156,25 +141,11 @@ contract MerchantRegistryTest is Test {
         vm.startPrank(merchant1);
 
         // Register first time
-        registry.registerMerchant(
-            merchant1,
-            usdcToken,
-            MerchantRegistry.PayoutMode.USDC_ONLY,
-            0,
-            500,
-            feeRecipient
-        );
+        registry.registerMerchant(merchant1, usdcToken, MerchantRegistry.PayoutMode.USDC_ONLY, 0, 500, feeRecipient);
 
         // Try to register again
         vm.expectRevert("already registered");
-        registry.registerMerchant(
-            merchant1,
-            usdtToken,
-            MerchantRegistry.PayoutMode.SHARES_ONLY,
-            0,
-            300,
-            feeRecipient
-        );
+        registry.registerMerchant(merchant1, usdtToken, MerchantRegistry.PayoutMode.SHARES_ONLY, 0, 300, feeRecipient);
 
         vm.stopPrank();
     }
@@ -211,14 +182,7 @@ contract MerchantRegistryTest is Test {
         vm.prank(merchant1);
 
         vm.expectRevert("payout token not allowed");
-        registry.registerMerchant(
-            merchant1,
-            invalidToken,
-            MerchantRegistry.PayoutMode.USDC_ONLY,
-            0,
-            500,
-            feeRecipient
-        );
+        registry.registerMerchant(merchant1, invalidToken, MerchantRegistry.PayoutMode.USDC_ONLY, 0, 500, feeRecipient);
     }
 
     function test_RegisterMerchant_EdgeCases() public {
@@ -234,9 +198,7 @@ contract MerchantRegistryTest is Test {
             address(0) // No fee recipient
         );
 
-        MerchantRegistry.MerchantConfig memory config = registry.getConfig(
-            merchant1
-        );
+        MerchantRegistry.MerchantConfig memory config = registry.getConfig(merchant1);
         assertEq(config.protocolFeeBps, 0);
         assertEq(config.feeRecipient, address(0));
     }
@@ -254,9 +216,7 @@ contract MerchantRegistryTest is Test {
             feeRecipient
         );
 
-        MerchantRegistry.MerchantConfig memory config = registry.getConfig(
-            merchant1
-        );
+        MerchantRegistry.MerchantConfig memory config = registry.getConfig(merchant1);
         assertEq(config.splitBps, 10000);
         assertEq(config.protocolFeeBps, 1000);
     }
@@ -265,14 +225,7 @@ contract MerchantRegistryTest is Test {
     function test_UpdateConfig_Success() public {
         // First register merchant
         vm.prank(merchant1);
-        registry.registerMerchant(
-            merchant1,
-            usdcToken,
-            MerchantRegistry.PayoutMode.USDC_ONLY,
-            0,
-            500,
-            feeRecipient
-        );
+        registry.registerMerchant(merchant1, usdcToken, MerchantRegistry.PayoutMode.USDC_ONLY, 0, 500, feeRecipient);
 
         // Update config
         vm.prank(merchant1);
@@ -288,9 +241,7 @@ contract MerchantRegistryTest is Test {
             address(0) // Change fee recipient
         );
 
-        MerchantRegistry.MerchantConfig memory config = registry.getConfig(
-            merchant1
-        );
+        MerchantRegistry.MerchantConfig memory config = registry.getConfig(merchant1);
         assertEq(config.payoutToken, usdtToken);
         assertEq(uint8(config.mode), uint8(MerchantRegistry.PayoutMode.SPLIT));
         assertEq(config.splitBps, 6000);
@@ -302,52 +253,24 @@ contract MerchantRegistryTest is Test {
     function test_UpdateConfig_NotRegistered() public {
         vm.prank(merchant1);
         vm.expectRevert("not registered");
-        registry.updateConfig(
-            merchant1,
-            usdcToken,
-            MerchantRegistry.PayoutMode.USDC_ONLY,
-            0,
-            500,
-            feeRecipient
-        );
+        registry.updateConfig(merchant1, usdcToken, MerchantRegistry.PayoutMode.USDC_ONLY, 0, 500, feeRecipient);
     }
 
     function test_UpdateConfig_NotMerchant() public {
         // Register merchant1
         vm.prank(merchant1);
-        registry.registerMerchant(
-            merchant1,
-            usdcToken,
-            MerchantRegistry.PayoutMode.USDC_ONLY,
-            0,
-            500,
-            feeRecipient
-        );
+        registry.registerMerchant(merchant1, usdcToken, MerchantRegistry.PayoutMode.USDC_ONLY, 0, 500, feeRecipient);
 
         // Try to update as different address
         vm.prank(merchant2);
         vm.expectRevert("not merchant");
-        registry.updateConfig(
-            merchant1,
-            usdtToken,
-            MerchantRegistry.PayoutMode.SHARES_ONLY,
-            0,
-            300,
-            feeRecipient
-        );
+        registry.updateConfig(merchant1, usdtToken, MerchantRegistry.PayoutMode.SHARES_ONLY, 0, 300, feeRecipient);
     }
 
     function test_UpdateConfig_InvalidSplitBps() public {
         // Register merchant
         vm.prank(merchant1);
-        registry.registerMerchant(
-            merchant1,
-            usdcToken,
-            MerchantRegistry.PayoutMode.USDC_ONLY,
-            0,
-            500,
-            feeRecipient
-        );
+        registry.registerMerchant(merchant1, usdcToken, MerchantRegistry.PayoutMode.USDC_ONLY, 0, 500, feeRecipient);
 
         // Try to update with invalid split
         vm.prank(merchant1);
@@ -365,14 +288,7 @@ contract MerchantRegistryTest is Test {
     function test_UpdateConfig_ProtocolFeeTooHigh() public {
         // Register merchant
         vm.prank(merchant1);
-        registry.registerMerchant(
-            merchant1,
-            usdcToken,
-            MerchantRegistry.PayoutMode.USDC_ONLY,
-            0,
-            500,
-            feeRecipient
-        );
+        registry.registerMerchant(merchant1, usdcToken, MerchantRegistry.PayoutMode.USDC_ONLY, 0, 500, feeRecipient);
 
         // Try to update with invalid fee
         vm.prank(merchant1);
@@ -390,33 +306,17 @@ contract MerchantRegistryTest is Test {
     function test_UpdateConfig_TokenNotAllowed() public {
         // Register merchant
         vm.prank(merchant1);
-        registry.registerMerchant(
-            merchant1,
-            usdcToken,
-            MerchantRegistry.PayoutMode.USDC_ONLY,
-            0,
-            500,
-            feeRecipient
-        );
+        registry.registerMerchant(merchant1, usdcToken, MerchantRegistry.PayoutMode.USDC_ONLY, 0, 500, feeRecipient);
 
         // Try to update with invalid token
         vm.prank(merchant1);
         vm.expectRevert("payout token not allowed");
-        registry.updateConfig(
-            merchant1,
-            invalidToken,
-            MerchantRegistry.PayoutMode.USDC_ONLY,
-            0,
-            500,
-            feeRecipient
-        );
+        registry.updateConfig(merchant1, invalidToken, MerchantRegistry.PayoutMode.USDC_ONLY, 0, 500, feeRecipient);
     }
 
     // === Getter Tests ===
     function test_GetConfig_NonExistentMerchant() public {
-        MerchantRegistry.MerchantConfig memory config = registry.getConfig(
-            merchant1
-        );
+        MerchantRegistry.MerchantConfig memory config = registry.getConfig(merchant1);
         assertFalse(config.exists);
         assertEq(config.merchantPayout, address(0));
         assertEq(config.payoutToken, address(0));
@@ -436,25 +336,11 @@ contract MerchantRegistryTest is Test {
     function test_CompleteWorkflow() public {
         // 1. Register merchant1 with USDC_ONLY
         vm.prank(merchant1);
-        registry.registerMerchant(
-            merchant1,
-            usdcToken,
-            MerchantRegistry.PayoutMode.USDC_ONLY,
-            0,
-            500,
-            feeRecipient
-        );
+        registry.registerMerchant(merchant1, usdcToken, MerchantRegistry.PayoutMode.USDC_ONLY, 0, 500, feeRecipient);
 
         // 2. Register merchant2 with SPLIT
         vm.prank(merchant2);
-        registry.registerMerchant(
-            merchant2,
-            usdtToken,
-            MerchantRegistry.PayoutMode.SPLIT,
-            7500,
-            300,
-            feeRecipient
-        );
+        registry.registerMerchant(merchant2, usdtToken, MerchantRegistry.PayoutMode.SPLIT, 7500, 300, feeRecipient);
 
         // 3. Update merchant1 to SHARES_ONLY
         vm.prank(merchant1);
@@ -468,20 +354,13 @@ contract MerchantRegistryTest is Test {
         );
 
         // 4. Verify final states
-        MerchantRegistry.MerchantConfig memory config1 = registry.getConfig(
-            merchant1
-        );
+        MerchantRegistry.MerchantConfig memory config1 = registry.getConfig(merchant1);
         assertEq(config1.payoutToken, daiToken);
-        assertEq(
-            uint8(config1.mode),
-            uint8(MerchantRegistry.PayoutMode.SHARES_ONLY)
-        );
+        assertEq(uint8(config1.mode), uint8(MerchantRegistry.PayoutMode.SHARES_ONLY));
         assertEq(config1.protocolFeeBps, 200);
         assertEq(config1.feeRecipient, merchant1);
 
-        MerchantRegistry.MerchantConfig memory config2 = registry.getConfig(
-            merchant2
-        );
+        MerchantRegistry.MerchantConfig memory config2 = registry.getConfig(merchant2);
         assertEq(config2.payoutToken, usdtToken);
         assertEq(uint8(config2.mode), uint8(MerchantRegistry.PayoutMode.SPLIT));
         assertEq(config2.splitBps, 7500);
@@ -495,7 +374,7 @@ contract MerchantRegistryTest is Test {
         tokens[2] = makeAddr("token3");
 
         // Add tokens
-        for (uint i = 0; i < tokens.length; i++) {
+        for (uint256 i = 0; i < tokens.length; i++) {
             registry.setAllowedToken(tokens[i], true);
             assertTrue(registry.isAllowedToken(tokens[i]));
         }
@@ -520,9 +399,7 @@ contract MerchantRegistryTest is Test {
             feeRecipient
         );
 
-        MerchantRegistry.MerchantConfig memory config = registry.getConfig(
-            address(0)
-        );
+        MerchantRegistry.MerchantConfig memory config = registry.getConfig(address(0));
         assertTrue(config.exists);
         assertEq(config.merchantPayout, merchant1);
     }
@@ -539,9 +416,7 @@ contract MerchantRegistryTest is Test {
             merchant1 // Self as fee recipient
         );
 
-        MerchantRegistry.MerchantConfig memory config = registry.getConfig(
-            merchant1
-        );
+        MerchantRegistry.MerchantConfig memory config = registry.getConfig(merchant1);
         assertEq(config.merchantPayout, merchant1);
         assertEq(config.feeRecipient, merchant1);
     }
@@ -551,39 +426,21 @@ contract MerchantRegistryTest is Test {
         vm.assume(splitBps <= 10000);
 
         vm.prank(merchant1);
-        registry.registerMerchant(
-            merchant1,
-            usdcToken,
-            MerchantRegistry.PayoutMode.SPLIT,
-            splitBps,
-            500,
-            feeRecipient
-        );
+        registry.registerMerchant(merchant1, usdcToken, MerchantRegistry.PayoutMode.SPLIT, splitBps, 500, feeRecipient);
 
-        MerchantRegistry.MerchantConfig memory config = registry.getConfig(
-            merchant1
-        );
+        MerchantRegistry.MerchantConfig memory config = registry.getConfig(merchant1);
         assertEq(config.splitBps, splitBps);
     }
 
-    function testFuzz_RegisterMerchant_ValidProtocolFee(
-        uint16 protocolFeeBps
-    ) public {
+    function testFuzz_RegisterMerchant_ValidProtocolFee(uint16 protocolFeeBps) public {
         vm.assume(protocolFeeBps <= 1000);
 
         vm.prank(merchant1);
         registry.registerMerchant(
-            merchant1,
-            usdcToken,
-            MerchantRegistry.PayoutMode.USDC_ONLY,
-            0,
-            protocolFeeBps,
-            feeRecipient
+            merchant1, usdcToken, MerchantRegistry.PayoutMode.USDC_ONLY, 0, protocolFeeBps, feeRecipient
         );
 
-        MerchantRegistry.MerchantConfig memory config = registry.getConfig(
-            merchant1
-        );
+        MerchantRegistry.MerchantConfig memory config = registry.getConfig(merchant1);
         assertEq(config.protocolFeeBps, protocolFeeBps);
     }
 
@@ -592,30 +449,16 @@ contract MerchantRegistryTest is Test {
 
         vm.prank(merchant1);
         vm.expectRevert("bad split");
-        registry.registerMerchant(
-            merchant1,
-            usdcToken,
-            MerchantRegistry.PayoutMode.SPLIT,
-            splitBps,
-            500,
-            feeRecipient
-        );
+        registry.registerMerchant(merchant1, usdcToken, MerchantRegistry.PayoutMode.SPLIT, splitBps, 500, feeRecipient);
     }
 
-    function testFuzz_RegisterMerchant_InvalidProtocolFee(
-        uint16 protocolFeeBps
-    ) public {
+    function testFuzz_RegisterMerchant_InvalidProtocolFee(uint16 protocolFeeBps) public {
         vm.assume(protocolFeeBps > 1000);
 
         vm.prank(merchant1);
         vm.expectRevert("fee too high");
         registry.registerMerchant(
-            merchant1,
-            usdcToken,
-            MerchantRegistry.PayoutMode.USDC_ONLY,
-            0,
-            protocolFeeBps,
-            feeRecipient
+            merchant1, usdcToken, MerchantRegistry.PayoutMode.USDC_ONLY, 0, protocolFeeBps, feeRecipient
         );
     }
 }
